+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+### pyBOAT 0.8.2
+
+- Added time averaging of Wavelet spectra <-> Fourier estimates
+- Added Fourier distribution for ensembles
+- Reworked FFT visualizations
+
 ### pyBOAT 0.8.17
 
 - Fixed crashes during batch analysis with thresholded ridges